refactor: 필드 결합도 개선
MemberProfileController-AddressApiService 느슨한 결합으로 개선
